now deserializez in the getter of the settings class and moved the settings class to a settings folder
